requirements.txt
gradio
google-generativeai
PyYAML
python-dotenv